[Chore] Blueprint url prefix
