+++ versions/v1/business.py
@@ -3,8 +3,9 @@ this way we can safely use decorator app.route()
 """
 from flask_jsonpify import jsonify
 from flask import request, Blueprint
+from versions.utils import find_business_by_id
 from versions import business_instance, login_required
-from versions.utils import find_business_by_id, check_if_biz_name_taken, biz_name_regex
+from versions.utils import check_if_biz_name_taken, biz_name_regex
 
 mod = Blueprint('business', __name__)
